Moves pages into new folder
